// apps/interactive-doc/src/components/ui/trigger/interface.ts
import { RefObject, ReactNode, SyntheticEvent } from "react"

export type TriggerColorScheme =
  | string
  | "white"
  | "gray"
  | "grayBlue"
  | "blackAlpha"
  | "red"
  | "orange"
  | "yellow"
  | "green"
  | "blue"
  | "cyan"
  | "purple"
  | "techPink"
  | "techPurple"

export type TriggerPosition =
  | "top"
  | "top-start"
  | "top-end"
  | "bottom"
  | "bottom-start"
  | "bottom-end"
  | "left"
  | "left-start"
  | "left-end"
  | "right"
  | "right-start"
  | "right-end"

export type TriggerTrigger = "hover" | "click" | "focus" | "contextmenu"

export interface TriggerContext {
  renderInBody?: boolean
  zIndex?: number
  children?: ReactNode
}

export interface TriggerRefHandler {
  rerenderPosition: () => void
}

export interface TriggerProps extends TriggerContext {
  [key: string]: any
  onBlur?: (e: SyntheticEvent) => void
  onFocus?: (e: SyntheticEvent) => void
  triggerRef?: RefObject<TriggerRefHandler | undefined>
  children?: ReactNode
  colorScheme?: TriggerColorScheme
  inline?: boolean
  closeWhenScroll?: boolean
  clickOutsideToClose?: boolean
  withoutPadding?: boolean
  withoutShadow?: boolean
  withoutOffset?: boolean
  trigger?: TriggerTrigger
  closeOnInnerClick?: boolean
  content?: string | ReactNode
  position?: TriggerPosition
  showArrow?: boolean
  closeDelay?: number
  openDelay?: number
  autoFitPosition?: boolean
  closeOnClick?: boolean
  defaultPopupVisible?: boolean
  autoAlignPopupWidth?: boolean
  popupVisible?: boolean
  disabled?: boolean
  onVisibleChange?: (visible: boolean) => void
  alignPoint?: boolean
  popupContainer?: HTMLElement | RefObject<HTMLElement>
}